Revert systemd-udev change (MACAddressPolicy)  that disables networking from pods
```
systemd-udev - Linux distributions that install the systemd-udev package create /usr/lib/systemd/network/99-default.link with Link.MACAddressPolicy set to persistent. This policy may cause the MAC address assigned to the host veth interface for a pod to change after the interface is moved to the host network namespace. The CNI plugin installs a static ARP binding for the default gateway in the pod network namespace pointing to the host veth MAC, so the MAC changing leads to pod connectivity issues. The workaround for this issue is to set MACAddressPolicy=none, as shown here. This issue is known to affect Ubuntu 22.04+, and long-term solutions are being evaluated.
```

fixes problem where there is no networking from pods! first symptom is that coredns crash loops as it can't talk to apiserver

Signed-off-by: Davanum Srinivas <davanum@gmail.com>